diet(phase10): split tools/lsp.rs (1257 LOC) into 7-file module
Mechanical split — covers the architecture-audit P6 normalization
(tools/ flat ↔ subdir consistency). External callers (tools/mod.rs
dispatch table, report_verifier, workflows) keep their `lsp::<fn>`
imports working via mod.rs re-exports.

tools/lsp/
- mod.rs (15)         re-exports
- shared.rs (115)     resolve_path_argument, insert_response_annotations,
                      lsp_install_hint, enhance_lsp_error,
                      attach_alias_warning, language_name_for_path
- diagnostics.rs (268) get_file_diagnostics + DiagnosticSourceContext +
                       pyright_* + suppression helpers
- references.rs (602) find_referencing_symbols (~427 LOC) +
                      compact_text_references + TS evidence helpers
- symbols.rs (126)    search_workspace_symbols + get_type_hierarchy
- rename.rs (159)     plan_symbol_rename + resolve_symbol_target +
                      resolve_symbol_position
- status.rs (26)      check_lsp_status + get_lsp_recipe

Surfaces one already-present dead pub fn (`check_lsp_status` —
no dispatch wiring, no tools.toml entry) — same pattern as phase7's
removed tools. Cleaned up in phase11.

Gates:
- cargo check --workspace --features http,semantic
- cargo test -p codelens-mcp --bin codelens-mcp (593 passed, 0 failed)
- cargo fmt --all -- --check
