+++ crates/codelens-mcp/src/tools/lsp/references.rs
@@ -1,210 +1,14 @@
-use super::{
+use super::super::{
     AppState, ToolResult, default_lsp_command_for_path, optional_bool, optional_string,
-    optional_usize, parse_lsp_args, required_string, success_meta,
+    optional_usize, parse_lsp_args, success_meta,
 };
+use super::rename::resolve_symbol_position;
+use super::shared::{enhance_lsp_error, insert_response_annotations, resolve_path_argument};
 use crate::error::CodeLensError;
 use crate::protocol::BackendKind;
 use crate::tool_evidence::{meta_degraded, meta_for_backend};
-use codelens_engine::{
-    LspDiagnosticRequest, LspRenamePlanRequest, LspRequest, LspResolveTargetRequest,
-    LspTypeHierarchyRequest, LspWorkspaceSymbolRequest, check_lsp_status as core_check_lsp_status,
-    extract_word_at_position, find_referencing_symbols_via_text,
-    get_lsp_recipe as core_get_lsp_recipe, get_type_hierarchy_native,
-};
+use codelens_engine::{LspRequest, extract_word_at_position, find_referencing_symbols_via_text};
 use serde_json::{Value, json};
-use std::collections::HashSet;
-
-fn resolve_path_argument(
-    arguments: &serde_json::Value,
-) -> Result<(&str, Vec<Value>), CodeLensError> {
-    if let Some(path) = optional_string(arguments, "path") {
-        if let Some(alias @ ("file_path" | "relative_path")) =
-            optional_string(arguments, "_path_alias_source")
-        {
-            return Ok((path, vec![crate::tool_runtime::path_alias_warning(alias)]));
-        }
-        return Ok((path, Vec::new()));
-    }
-    for alias in ["file_path", "relative_path"] {
-        if let Some(path) = optional_string(arguments, alias) {
-            return Ok((path, vec![crate::tool_runtime::path_alias_warning(alias)]));
-        }
-    }
-    Err(CodeLensError::MissingParam("path".to_owned()))
-}
-
-fn insert_response_annotations(
-    payload: &mut Value,
-    unknown_args: &[String],
-    deprecation_warnings: &[Value],
-) {
-    let Some(map) = payload.as_object_mut() else {
-        return;
-    };
-    if !unknown_args.is_empty() {
-        map.insert("unknown_args".to_owned(), json!(unknown_args));
-    }
-    if !deprecation_warnings.is_empty() {
-        map.insert(
-            "deprecation_warnings".to_owned(),
-            json!(deprecation_warnings),
-        );
-    }
-}
-
-fn postprocess_lsp_diagnostics(
-    project: &codelens_engine::ProjectRoot,
-    file_path: &str,
-    diagnostics: Vec<codelens_engine::LspDiagnostic>,
-) -> (Vec<Value>, Vec<Value>) {
-    let source = DiagnosticSourceContext::load(project, file_path);
-    let mut visible = Vec::new();
-    let mut suppressed = Vec::new();
-
-    for diagnostic in diagnostics {
-        if let Some(reason) = pyright_source_suppression_reason(&source, &diagnostic) {
-            suppressed.push(json!({
-                "file_path": diagnostic.file_path,
-                "line": diagnostic.line,
-                "column": diagnostic.column,
-                "code": diagnostic.code,
-                "source": diagnostic.source,
-                "message": diagnostic.message,
-                "suppression": reason,
-            }));
-            continue;
-        }
-
-        let mut value = serde_json::to_value(&diagnostic).unwrap_or_else(|_| json!({}));
-        if is_optional_import_diagnostic(&source, &diagnostic) {
-            value["classification"] = json!("optional_dependency_import");
-            value["actionability"] = json!("environmental_optional_dependency");
-            value["recommended_action"] = json!(
-                "Do not patch source solely for this diagnostic; install the optional extra or treat it as non-blocking when the import is guarded by ImportError."
-            );
-        }
-        visible.push(value);
-    }
-
-    (visible, suppressed)
-}
-
-struct DiagnosticSourceContext {
-    lines: Vec<String>,
-    disabled_pyright_rules: HashSet<String>,
-}
-
-impl DiagnosticSourceContext {
-    fn load(project: &codelens_engine::ProjectRoot, file_path: &str) -> Self {
-        let source = project
-            .resolve(file_path)
-            .ok()
-            .and_then(|path| std::fs::read_to_string(path).ok())
-            .unwrap_or_default();
-        let lines: Vec<String> = source.lines().map(ToOwned::to_owned).collect();
-        let disabled_pyright_rules = lines
-            .iter()
-            .take(10)
-            .flat_map(|line| disabled_pyright_rules_on_line(line))
-            .collect();
-        Self {
-            lines,
-            disabled_pyright_rules,
-        }
-    }
-
-    fn line(&self, one_based_line: usize) -> Option<&str> {
-        self.lines
-            .get(one_based_line.saturating_sub(1))
-            .map(String::as_str)
-    }
-}
-
-fn pyright_source_suppression_reason(
-    source: &DiagnosticSourceContext,
-    diagnostic: &codelens_engine::LspDiagnostic,
-) -> Option<&'static str> {
-    let code = diagnostic.code.as_deref()?;
-    if !is_pyright_diagnostic(diagnostic, code) {
-        return None;
-    }
-    if source.disabled_pyright_rules.contains(code) {
-        return Some("file_pyright_rule_disabled");
-    }
-    let line = source.line(diagnostic.line)?;
-    if line_suppresses_pyright_rule(line, code) {
-        return Some("line_pyright_ignore");
-    }
-    None
-}
-
-fn is_pyright_diagnostic(diagnostic: &codelens_engine::LspDiagnostic, code: &str) -> bool {
-    diagnostic.source.as_deref() == Some("pyright") || code.starts_with("report")
-}
-
-fn disabled_pyright_rules_on_line(line: &str) -> Vec<String> {
-    let normalized = normalize_pyright_directive(line);
-    let Some((_, directive)) = normalized.split_once("#pyright:") else {
-        return Vec::new();
-    };
-    directive
-        .split(',')
-        .filter_map(|entry| {
-            let (rule, value) = entry.split_once('=')?;
-            (value == "false").then(|| rule.to_owned())
-        })
-        .collect()
-}
-
-fn line_suppresses_pyright_rule(line: &str, code: &str) -> bool {
-    let normalized = normalize_pyright_directive(line);
-    if normalized.contains("#type:ignore") {
-        return true;
-    }
-    if !normalized.contains("#pyright:ignore") {
-        return false;
-    }
-    !normalized.contains('[') || normalized.contains(code)
-}
-
-fn normalize_pyright_directive(line: &str) -> String {
-    line.chars().filter(|ch| !ch.is_whitespace()).collect()
-}
-
-fn is_optional_import_diagnostic(
-    source: &DiagnosticSourceContext,
-    diagnostic: &codelens_engine::LspDiagnostic,
-) -> bool {
-    let Some(code) = diagnostic.code.as_deref() else {
-        return false;
-    };
-    if !matches!(code, "reportMissingImports" | "reportMissingModuleSource") {
-        return false;
-    }
-    import_line_is_guarded_by_import_error(source, diagnostic.line)
-}
-
-fn import_line_is_guarded_by_import_error(
-    source: &DiagnosticSourceContext,
-    one_based_line: usize,
-) -> bool {
-    let Some(import_line) = source.line(one_based_line) else {
-        return false;
-    };
-    if !import_line.contains("import ") {
-        return false;
-    }
-    let line_index = one_based_line.saturating_sub(1);
-    let start = line_index.saturating_sub(4);
-    let end = (line_index + 8).min(source.lines.len().saturating_sub(1));
-    let has_try = source.lines[start..=line_index]
-        .iter()
-        .any(|line| line.trim_start().starts_with("try:"));
-    let has_import_error_handler = source.lines[line_index..=end]
-        .iter()
-        .any(|line| line.trim_start().starts_with("except ImportError"));
-    has_try && has_import_error_handler
-}
 
 fn compact_text_references(
     references: Vec<codelens_engine::TextReference>,
@@ -358,50 +162,6 @@ fn insert_structural_ts_evidence(
             "recommended_action": "inspect_structural_reference_evidence",
         }),
     );
-}
-
-fn lsp_install_hint(command: &str) -> &'static str {
-    match command {
-        "pyright" => "  pip install pyright",
-        "typescript-language-server" => "  npm i -g typescript-language-server typescript",
-        "rust-analyzer" => "  rustup component add rust-analyzer",
-        "gopls" => "  go install golang.org/x/tools/gopls@latest",
-        "clangd" => "  brew install llvm  (or apt install clangd)",
-        "jdtls" => "  See https://github.com/eclipse-jdtls/eclipse.jdt.ls",
-        "solargraph" => "  gem install solargraph",
-        "intelephense" => "  npm i -g intelephense",
-        "kotlin-language-server" => "  See https://github.com/fwcd/kotlin-language-server",
-        "metals" => "  cs install metals  (via Coursier)",
-        "sourcekit-lsp" => "  Included with Xcode / Swift toolchain",
-        "csharp-ls" => "  dotnet tool install -g csharp-ls",
-        "dart" => "  dart pub global activate dart_language_server",
-        // Phase 6a languages
-        "lua-language-server" => "  brew install lua-language-server",
-        "zls" => "  brew install zls",
-        "nextls" => "  mix escript.install hex next_ls",
-        "haskell-language-server-wrapper" => "  ghcup install hls",
-        "ocamllsp" => "  opam install ocaml-lsp-server",
-        "erlang_ls" => "  brew install erlang_ls",
-        "bash-language-server" => "  npm i -g bash-language-server",
-        _ => "  Check your package manager for the LSP server binary",
-    }
-}
-
-fn enhance_lsp_error(err: anyhow::Error, command: &str) -> CodeLensError {
-    let msg = err.to_string();
-    if msg.contains("No such file") || msg.contains("not found") || msg.contains("spawn") {
-        CodeLensError::LspNotAttached(format!(
-            "LSP server '{command}' not found. Install it:\n{}",
-            lsp_install_hint(command)
-        ))
-    } else if msg.contains("timed out") || msg.contains("timeout") {
-        CodeLensError::Timeout {
-            operation: format!("LSP {command}"),
-            elapsed_ms: 30_000,
-        }
-    } else {
-        CodeLensError::LspError(msg)
-    }
 }
 
 /// tree-sitter-first strategy:
@@ -839,419 +599,4 @@ pub fn find_referencing_symbols(state: &AppState, arguments: &serde_json::Value)
             (payload, meta)
         })?,
     )
-}
-
-/// Resolve a symbol name to its (line, column) position in a file via the symbol index.
-fn resolve_symbol_position(
-    state: &AppState,
-    symbol_name: &str,
-    file_path: &str,
-) -> Option<(usize, usize)> {
-    let symbols = state
-        .symbol_index()
-        .find_symbol(symbol_name, Some(file_path), false, true, 1)
-        .ok()?;
-    symbols.first().map(|s| (s.line, s.column))
-}
-
-pub fn get_file_diagnostics(state: &AppState, arguments: &serde_json::Value) -> ToolResult {
-    const KNOWN_ARGS: &[&str] = &[
-        "path",
-        "file_path",
-        "relative_path",
-        "command",
-        "args",
-        "max_results",
-    ];
-    let (file_path_arg, deprecation_warnings) = resolve_path_argument(arguments)?;
-    let file_path = file_path_arg.to_owned();
-    let max_results = optional_usize(arguments, "max_results", 200);
-    let unknown_args = crate::tool_runtime::collect_unknown_args(arguments, KNOWN_ARGS);
-
-    // Try SCIP diagnostics first (if available).
-    #[cfg(feature = "scip-backend")]
-    if let Some(backend) = state.scip() {
-        use codelens_engine::PreciseBackend as _;
-        if let Ok(scip_diags) = backend.diagnostics(&file_path)
-            && !scip_diags.is_empty()
-        {
-            let limited: Vec<_> = scip_diags.into_iter().take(max_results).collect();
-            let count = limited.len();
-            let diags_json: Vec<serde_json::Value> = limited
-                .iter()
-                .map(|d| {
-                    json!({
-                        "file_path": d.file_path,
-                        "line": d.line,
-                        "column": d.column,
-                        "severity": format!("{:?}", d.severity),
-                        "message": d.message,
-                        "source": "scip",
-                        "code": d.code,
-                    })
-                })
-                .collect();
-            let mut payload = json!({
-                "diagnostics": diags_json,
-                "count": count,
-                "backend": "scip",
-            });
-            insert_response_annotations(&mut payload, &unknown_args, &deprecation_warnings);
-            return Ok((payload, success_meta(BackendKind::Scip, 0.95)));
-        }
-    }
-
-    // Fall back to LSP diagnostics.
-    let command = optional_string(arguments, "command")
-        .map(ToOwned::to_owned)
-        .or_else(|| default_lsp_command_for_path(&file_path))
-        .ok_or_else(|| CodeLensError::LspError("no default LSP mapping for file".into()))?;
-    let args = parse_lsp_args(arguments, &command);
-
-    let command_ref = command.clone();
-    let request_file_path = file_path.clone();
-    state
-        .lsp_pool()
-        .get_diagnostics(&LspDiagnosticRequest {
-            command,
-            args,
-            file_path,
-            max_results,
-        })
-        .map_err(|e| enhance_lsp_error(e, &command_ref))
-        .map(|value| {
-            let (diagnostics, suppressed) =
-                postprocess_lsp_diagnostics(&state.project(), &request_file_path, value);
-            let mut payload = json!({
-                "diagnostics": diagnostics,
-                "count": diagnostics.len(),
-                "backend": "lsp",
-            });
-            if !suppressed.is_empty() {
-                payload["suppressed_diagnostics"] = json!(suppressed);
-                payload["suppressed_diagnostics_count"] = json!(
-                    payload["suppressed_diagnostics"]
-                        .as_array()
-                        .map_or(0, Vec::len)
-                );
-            }
-            insert_response_annotations(&mut payload, &unknown_args, &deprecation_warnings);
-            (payload, success_meta(BackendKind::Lsp, 0.9))
-        })
-}
-
-pub fn search_workspace_symbols(state: &AppState, arguments: &serde_json::Value) -> ToolResult {
-    let query = required_string(arguments, "query")?.to_owned();
-    // `command` is the LSP server binary (rust-analyzer / pyright / gopls …).
-    // When missing, point users at the non-LSP fuzzy fallback instead of the
-    // generic "Missing required parameter" error so CLI one-shot callers
-    // don't hit a dead end.
-    let Some(command) = optional_string(arguments, "command").map(ToOwned::to_owned) else {
-        return Err(CodeLensError::MissingParam(format!(
-            "command (LSP server binary, e.g. rust-analyzer/pyright). \
-             For LSP-free fuzzy search over `{query}`, call \
-             `bm25_symbol_search` (or `find_symbol` with an exact name)."
-        )));
-    };
-    let args = parse_lsp_args(arguments, &command);
-    let max_results = optional_usize(arguments, "max_results", 50);
-
-    let command_ref = command.clone();
-    state
-        .lsp_pool()
-        .search_workspace_symbols(&LspWorkspaceSymbolRequest {
-            command,
-            args,
-            query,
-            max_results,
-        })
-        .map_err(|e| enhance_lsp_error(e, &command_ref))
-        .map(|value| {
-            (
-                json!({ "symbols": value, "count": value.len() }),
-                success_meta(BackendKind::Lsp, 0.88),
-            )
-        })
-}
-
-pub fn get_type_hierarchy(state: &AppState, arguments: &serde_json::Value) -> ToolResult {
-    let query = arguments
-        .get("name_path")
-        .or_else(|| arguments.get("fully_qualified_name"))
-        .and_then(|v| v.as_str())
-        .ok_or_else(|| CodeLensError::MissingParam("name_path or fully_qualified_name".into()))?
-        .to_owned();
-    // #180: prefer canonical `path`; envelope normalises `relative_path → path`
-    // for this tool, so reading `path` first picks up either input shape.
-    let relative_path = optional_string(arguments, "path")
-        .or_else(|| optional_string(arguments, "relative_path"))
-        .map(ToOwned::to_owned);
-    let alias_used = optional_string(arguments, "_path_alias_source")
-        .filter(|s| *s == "relative_path")
-        .map(crate::tool_runtime::path_alias_warning);
-    let hierarchy_type = optional_string(arguments, "hierarchy_type")
-        .unwrap_or("both")
-        .to_owned();
-    let depth = optional_usize(arguments, "depth", 1);
-    let command = optional_string(arguments, "command")
-        .map(ToOwned::to_owned)
-        .or_else(|| {
-            relative_path
-                .as_deref()
-                .and_then(default_lsp_command_for_path)
-        });
-
-    if let Some(command) = command {
-        let args = parse_lsp_args(arguments, &command);
-        let lsp_result = state
-            .lsp_pool()
-            .get_type_hierarchy(&LspTypeHierarchyRequest {
-                command,
-                args,
-                query: query.clone(),
-                relative_path: relative_path.clone(),
-                hierarchy_type: hierarchy_type.clone(),
-                depth: if depth == 0 { 8 } else { depth },
-            });
-
-        match lsp_result {
-            Ok(value) => Ok((
-                attach_alias_warning(json!(value), alias_used.clone()),
-                meta_for_backend("lsp_pooled", 0.82),
-            )),
-            Err(_) => Ok(get_type_hierarchy_native(
-                &state.project(),
-                &query,
-                relative_path.as_deref(),
-                &hierarchy_type,
-                depth,
-            )
-            .map(|value| {
-                (
-                    attach_alias_warning(json!(value), alias_used.clone()),
-                    meta_degraded(
-                        "tree-sitter-native",
-                        0.80,
-                        "LSP failed, fell back to native",
-                    ),
-                )
-            })?),
-        }
-    } else {
-        Ok(get_type_hierarchy_native(
-            &state.project(),
-            &query,
-            relative_path.as_deref(),
-            &hierarchy_type,
-            depth,
-        )
-        .map(|value| {
-            (
-                attach_alias_warning(json!(value), alias_used),
-                meta_degraded("tree-sitter-native", 0.80, "no LSP command available"),
-            )
-        })?)
-    }
-}
-
-/// #180: append a single deprecation_warnings entry to a payload when the
-/// caller used a legacy path alias. Mirrors the per-file helper in
-/// filesystem.rs so all path-aliased tools emit the same shape.
-fn attach_alias_warning(mut payload: Value, warning: Option<Value>) -> Value {
-    if let Some(warning) = warning
-        && let Some(map) = payload.as_object_mut()
-    {
-        map.insert(
-            "deprecation_warnings".to_owned(),
-            Value::Array(vec![warning]),
-        );
-    }
-    payload
-}
-
-pub fn plan_symbol_rename(state: &AppState, arguments: &serde_json::Value) -> ToolResult {
-    const KNOWN_ARGS: &[&str] = &[
-        "path",
-        "file_path",
-        "line",
-        "column",
-        "new_name",
-        "command",
-        "args",
-    ];
-    let (file_path_arg, deprecation_warnings) = resolve_path_argument(arguments)?;
-    let unknown_args = crate::tool_runtime::collect_unknown_args(arguments, KNOWN_ARGS);
-    let file_path = file_path_arg.to_owned();
-    let line = arguments
-        .get("line")
-        .and_then(|v| v.as_u64())
-        .ok_or_else(|| CodeLensError::MissingParam("line".into()))? as usize;
-    let column = arguments
-        .get("column")
-        .and_then(|v| v.as_u64())
-        .ok_or_else(|| CodeLensError::MissingParam("column".into()))? as usize;
-    let new_name = optional_string(arguments, "new_name").map(ToOwned::to_owned);
-    let command = optional_string(arguments, "command")
-        .map(ToOwned::to_owned)
-        .or_else(|| default_lsp_command_for_path(&file_path))
-        .ok_or_else(|| CodeLensError::LspError("no default LSP mapping for file".into()))?;
-    let args = parse_lsp_args(arguments, &command);
-
-    let command_ref = command.clone();
-    state
-        .lsp_pool()
-        .get_rename_plan(&LspRenamePlanRequest {
-            command,
-            args,
-            file_path,
-            line,
-            column,
-            new_name,
-        })
-        .map_err(|e| enhance_lsp_error(e, &command_ref))
-        .map(|value| {
-            let mut payload = json!(value);
-            insert_response_annotations(&mut payload, &unknown_args, &deprecation_warnings);
-            (payload, success_meta(BackendKind::Lsp, 0.86))
-        })
-}
-
-pub fn resolve_symbol_target(state: &AppState, arguments: &serde_json::Value) -> ToolResult {
-    const KNOWN_ARGS: &[&str] = &[
-        "path",
-        "file_path",
-        "line",
-        "column",
-        "target",
-        "semantic_backend",
-        "command",
-        "args",
-        "max_results",
-    ];
-    let (file_path_arg, deprecation_warnings) = resolve_path_argument(arguments)?;
-    let unknown_args = crate::tool_runtime::collect_unknown_args(arguments, KNOWN_ARGS);
-    let file_path = file_path_arg.to_owned();
-    let line = arguments
-        .get("line")
-        .and_then(|v| v.as_u64())
-        .ok_or_else(|| CodeLensError::MissingParam("line".into()))? as usize;
-    let column = arguments
-        .get("column")
-        .and_then(|v| v.as_u64())
-        .ok_or_else(|| CodeLensError::MissingParam("column".into()))? as usize;
-    let target = optional_string(arguments, "target")
-        .unwrap_or("definition")
-        .to_owned();
-    let semantic_backend = optional_string(arguments, "semantic_backend").unwrap_or("lsp");
-    if semantic_backend != "lsp" {
-        return Err(CodeLensError::Validation(
-            "resolve_symbol_target currently supports semantic_backend=lsp only".into(),
-        ));
-    }
-    if !matches!(
-        target.as_str(),
-        "declaration" | "definition" | "implementation" | "type_definition"
-    ) {
-        return Err(CodeLensError::Validation(format!(
-            "unsupported resolve target `{target}`"
-        )));
-    }
-    let command = optional_string(arguments, "command")
-        .map(ToOwned::to_owned)
-        .or_else(|| default_lsp_command_for_path(&file_path))
-        .ok_or_else(|| CodeLensError::LspError("no default LSP mapping for file".into()))?;
-    let args = parse_lsp_args(arguments, &command);
-    let max_results = optional_usize(arguments, "max_results", 20);
-
-    let command_ref = command.clone();
-    state
-        .lsp_pool()
-        .resolve_symbol_target(&LspResolveTargetRequest {
-            command,
-            args,
-            file_path: file_path.clone(),
-            line,
-            column,
-            target: target.clone(),
-            max_results,
-        })
-        .map_err(|e| enhance_lsp_error(e, &command_ref))
-        .map(|targets| {
-            let method = match target.as_str() {
-                "declaration" => "textDocument/declaration",
-                "definition" => "textDocument/definition",
-                "implementation" => "textDocument/implementation",
-                "type_definition" => "textDocument/typeDefinition",
-                _ => "unknown",
-            };
-            let mut payload = json!({
-                "success": true,
-                "semantic_backend": "lsp",
-                "edit_authority": {
-                    "kind": "authoritative_lsp",
-                    "backend": "lsp",
-                    "operation": target,
-                    "language": language_name_for_path(&file_path),
-                    "methods": [method],
-                    "embedding_used": false,
-                    "search_used": false
-                },
-                "targets": targets,
-                "count": targets.len(),
-            });
-            insert_response_annotations(&mut payload, &unknown_args, &deprecation_warnings);
-            (payload, success_meta(BackendKind::Lsp, 0.95))
-        })
-}
-
-#[allow(dead_code)]
-pub fn check_lsp_status(_state: &AppState, _arguments: &serde_json::Value) -> ToolResult {
-    let statuses = core_check_lsp_status();
-    Ok((
-        json!({ "servers": statuses, "count": statuses.len() }),
-        success_meta(BackendKind::Lsp, 1.0),
-    ))
-}
-
-fn language_name_for_path(file_path: &str) -> &'static str {
-    match std::path::Path::new(file_path)
-        .extension()
-        .and_then(|ext| ext.to_str())
-        .unwrap_or_default()
-    {
-        "rs" => "rust",
-        "ts" | "tsx" => "typescript",
-        "js" | "jsx" | "mjs" | "cjs" => "javascript",
-        "java" => "java",
-        "py" => "python",
-        _ => "unknown",
-    }
-}
-
-pub fn get_lsp_recipe(_state: &AppState, arguments: &serde_json::Value) -> ToolResult {
-    let extension = required_string(arguments, "extension")?;
-    match core_get_lsp_recipe(extension) {
-        Some(recipe) => Ok((json!(recipe), success_meta(BackendKind::Lsp, 1.0))),
-        None => Err(CodeLensError::NotFound(format!(
-            "LSP recipe for extension: {extension}"
-        ))),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn disabled_pyright_rules_on_line_parses_multiple_rules() {
-        assert_eq!(
-            disabled_pyright_rules_on_line(
-                "# pyright: reportMissingImports=false, reportCallIssue=false"
-            ),
-            vec![
-                "reportMissingImports".to_owned(),
-                "reportCallIssue".to_owned()
-            ]
-        );
-    }
 }
